Create README.md
Tslot bench organizer system I designed for my small work bench, It is high density, it suspends the rail from the wall and allows mounting parts on all 4 sides. The mounts are made to hold the rail on top of bottom only, so that an led strip (11mm?) can be ran through it. It is also made to be maintained, as the entire thing can be removed from the wall without removing the mounts.

You can run an led strip on any of 3 sides ( which ever sided you are not using to mount ), I recommend using captive slide in nuts for the mounts, and not the twist insert ones. I will try to add more models and details on hardware when I have time. I think I uses steel washers in side the mount holes to distribute the force when screwing them in.